+++ backend/express/expressServer.js
@@ -1,6 +1,6 @@
 import express from "express";
-import router from "./router/router.js";
-import auth from "./router/authRouter.js";
+import router from "./src/router/router.js";
+import auth from "./src/router/authRouter.js";
 
 const port = 4500;
 
@@ -14,3 +14,5 @@ app.use("/api", router);
 app.listen(port, () => {
   console.log("Server started, listening on port: " + port);
 });
+
+console.log(process.env.URI);
