+++ backend/express/src/router/router.js
@@ -1,4 +1,6 @@
 import express from "express";
+import dotenv from "dotenv";
+dotenv.config();
 
 const router = express.Router();
